Moved mode and submode enums to separate files for other modules to access them easily
